processing_backend/requirements/core.requirements.txt
pymongo~=4.7.2
sentence-transformers~=2.7.0
python-dotenv~=1.0.1
pillow~=10.3.0
minio~=7.2.7
pydantic~=2.7.3
fastapi~=0.111.0
uvicorn~=0.30.1
